move fall_asleep() to character.cpp
