+++ src/character.cpp
@@ -55,7 +55,10 @@
 #include "rng.h"
 #include "stomach.h"
 #include "ui.h"
+#include "veh_type.h"
+#include "vehicle.h"
 #include "vitamin.h"
+#include "vpart_position.h"
 
 static const bionic_id bio_ads( "bio_ads" );
 static const bionic_id bio_armor_arms( "bio_armor_arms" );
@@ -5107,4 +5110,190 @@ double Character::footwear_factor() const
         ret += .5;
     }
     return ret;
+}
+
+void Character::start_hauling()
+{
+    add_msg( _( "You start hauling items along the ground." ) );
+    if( is_armed() ) {
+        add_msg( m_warning, _( "Your hands are not free, which makes hauling slower." ) );
+    }
+    hauling = true;
+}
+
+void Character::stop_hauling()
+{
+    add_msg( _( "You stop hauling items." ) );
+    hauling = false;
+    if( has_activity( activity_id( "ACT_MOVE_ITEMS" ) ) ) {
+        cancel_activity();
+    }
+}
+
+bool Character::is_hauling() const
+{
+    return hauling;
+}
+
+void Character::assign_activity( const activity_id &type, int moves, int index, int pos,
+                                 const std::string &name )
+{
+    assign_activity( player_activity( type, moves, index, pos, name ) );
+}
+
+void Character::assign_activity( const player_activity &act, bool allow_resume )
+{
+    if( allow_resume && !backlog.empty() && backlog.front().can_resume_with( act, *this ) ) {
+        add_msg_if_player( _( "You resume your task." ) );
+        activity = backlog.front();
+        backlog.pop_front();
+    } else {
+        if( activity ) {
+            backlog.push_front( activity );
+        }
+
+        activity = act;
+    }
+
+    if( activity.rooted() ) {
+        rooted_message();
+    }
+    if( is_npc() ) {
+        npc *guy = dynamic_cast<npc *>( this );
+        guy->set_attitude( NPCATT_ACTIVITY );
+        guy->set_mission( NPC_MISSION_ACTIVITY );
+        guy->current_activity_id = activity.id();
+    }
+}
+
+bool Character::has_activity( const activity_id &type ) const
+{
+    return activity.id() == type;
+}
+
+bool Character::has_activity( const std::vector<activity_id> &types ) const
+{
+    return std::find( types.begin(), types.end(), activity.id() ) != types.end();
+}
+
+void Character::cancel_activity()
+{
+    if( has_activity( activity_id( "ACT_MOVE_ITEMS" ) ) && is_hauling() ) {
+        stop_hauling();
+    }
+    if( has_activity( activity_id( "ACT_TRY_SLEEP" ) ) ) {
+        remove_value( "sleep_query" );
+    }
+    // Clear any backlog items that aren't auto-resume.
+    for( auto backlog_item = backlog.begin(); backlog_item != backlog.end(); ) {
+        if( backlog_item->auto_resume ) {
+            backlog_item++;
+        } else {
+            backlog_item = backlog.erase( backlog_item );
+        }
+    }
+    if( activity && activity.is_suspendable() ) {
+        backlog.push_front( activity );
+    }
+    sfx::end_activity_sounds(); // kill activity sounds when canceled
+    activity = player_activity();
+}
+
+void Character::resume_backlog_activity()
+{
+    if( !backlog.empty() && backlog.front().auto_resume ) {
+        activity = backlog.front();
+        backlog.pop_front();
+    }
+}
+
+void Character::fall_asleep()
+{
+    // Communicate to the player that he is using items on the floor
+    std::string item_name = is_snuggling();
+    if( item_name == "many" ) {
+        if( one_in( 15 ) ) {
+            add_msg_if_player( _( "You nestle your pile of clothes for warmth." ) );
+        } else {
+            add_msg_if_player( _( "You use your pile of clothes for warmth." ) );
+        }
+    } else if( item_name != "nothing" ) {
+        if( one_in( 15 ) ) {
+            add_msg_if_player( _( "You snuggle your %s to keep warm." ), item_name );
+        } else {
+            add_msg_if_player( _( "You use your %s to keep warm." ), item_name );
+        }
+    }
+    if( has_active_mutation( trait_id( "HIBERNATE" ) ) &&
+        get_kcal_percent() > 0.8f ) {
+        if( is_avatar() ) {
+            g->memorial().add( pgettext( "memorial_male", "Entered hibernation." ),
+                               pgettext( "memorial_female", "Entered hibernation." ) );
+        }
+        // some days worth of round-the-clock Snooze.  Cata seasons default to 91 days.
+        fall_asleep( 10_days );
+        // If you're not fatigued enough for 10 days, you won't sleep the whole thing.
+        // In practice, the fatigue from filling the tank from (no msg) to Time For Bed
+        // will last about 8 days.
+    }
+
+    fall_asleep( 10_hours ); // default max sleep time.
+}
+
+void Character::fall_asleep( const time_duration &duration )
+{
+    if( activity ) {
+        if( activity.id() == "ACT_TRY_SLEEP" ) {
+            activity.set_to_null();
+        } else {
+            cancel_activity();
+        }
+    }
+    add_effect( effect_sleep, duration );
+}
+
+std::string Character::is_snuggling() const
+{
+    auto begin = g->m.i_at( pos() ).begin();
+    auto end = g->m.i_at( pos() ).end();
+
+    if( in_vehicle ) {
+        if( const cata::optional<vpart_reference> vp = g->m.veh_at( pos() ).part_with_feature( VPFLAG_CARGO,
+                false ) ) {
+            vehicle *const veh = &vp->vehicle();
+            const int cargo = vp->part_index();
+            if( !veh->get_items( cargo ).empty() ) {
+                begin = veh->get_items( cargo ).begin();
+                end = veh->get_items( cargo ).end();
+            }
+        }
+    }
+    const item *floor_armor = nullptr;
+    int ticker = 0;
+
+    // If there are no items on the floor, return nothing
+    if( begin == end ) {
+        return "nothing";
+    }
+
+    for( auto candidate = begin; candidate != end; ++candidate ) {
+        if( !candidate->is_armor() ) {
+            continue;
+        } else if( candidate->volume() > 250_ml && candidate->get_warmth() > 0 &&
+                   ( candidate->covers( bp_torso ) || candidate->covers( bp_leg_l ) ||
+                     candidate->covers( bp_leg_r ) ) ) {
+            floor_armor = &*candidate;
+            ticker++;
+        }
+    }
+
+    if( ticker == 0 ) {
+        return "nothing";
+    } else if( ticker == 1 ) {
+        return floor_armor->type_name();
+    } else if( ticker > 1 ) {
+        return "many";
+    }
+
+    return "nothing";
 }
